+++ src/Facades/LaravelFiles.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Facade;
 
 /**
  * @method static \Vhar\LaravelFiles\Models\File upload(\Illuminate\Http\UploadedFile $file, ?string $disk = null, string $directory = '')
+ * @method static \Vhar\LaravelFiles\Models\File replace(\Vhar\LaravelFiles\Models\File $file, \Illuminate\Http\UploadedFile $uploadedFile)
  * @method static void attach(\Illuminate\Database\Eloquent\Model $model, \Vhar\LaravelFiles\Models\File $file, string $collection = 'default', int $sort = 0)
  * @method static void detach(\Illuminate\Database\Eloquent\Model $model, \Vhar\LaravelFiles\Models\File $file, string $collection = 'default')
  * @method static bool exists(\Vhar\LaravelFiles\Models\File $file)